Add name

+++ src/main/java/FormDate.java
@@ -1,6 +1,6 @@
 public class FormDate {
     public int day;
     public int month;
-    public int year;
+    public long year;
 
 }
